pyproject.toml
[project]
name = "mrs-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "backoff>=2.2.1",
    "dotenv>=0.9.9",
    "openai>=2.29.0",
    "pyyaml>=6.0.3",
    "ur-rtde>=1.6.3",
]
